% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/didint.r
\name{coef.DiDIntObj}
\alias{coef.DiDIntObj}
\title{Extract coefficients from \code{DiDIntObj}}
\usage{
\method{coef}{DiDIntObj}(object, level = c("agg", "sub"), ...)
}
\arguments{
\item{object}{A \code{DiDIntObj} object}

\item{level}{Specify either \code{"agg"} or \code{"sub"} to view the aggregate
or sub-aggregate results.}

\item{...}{other arguments}
}
\value{
A data frame of coefficient estimates
}
\description{
Extract coefficients from \code{DiDIntObj}
}
