% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/didint.r
\name{create_didint_object}
\alias{create_didint_object}
\title{DiDIntObj}
\usage{
create_didint_object(result, ccc, weighting, agg)
}
\arguments{
\item{result}{DataFrame of results from DiDInt.jl}

\item{ccc}{The specified CCC variation of DID-INT that was used.}

\item{weighting}{The weighting method that was used.}

\item{agg}{The aggregation method that was used.}
}
\value{
DiDIntObj with class "DiDIntObj"
}
\description{
Objects of this class store all of the results for a given
set of aggregation, weighting, and ccc options.
}
