% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/didint_plot.r
\name{didint_plot}
\alias{didint_plot}
\title{Make event study or parallel trends plots}
\usage{
didint_plot(
  outcome,
  state,
  time,
  data,
  gvar = NULL,
  treated_states = NULL,
  treatment_times = NULL,
  date_format = NULL,
  covariates = NULL,
  ref = NULL,
  ccc = "all",
  event = FALSE,
  weights = TRUE,
  ci = 0.95,
  freq = NULL,
  freq_multiplier = 1,
  start_date = NULL,
  end_date = NULL,
  hc = "hc3"
)
}
\arguments{
\item{outcome}{A string giving the column name of the outcome variable.}

\item{state}{A string giving the column identifying states. The state column
should be a character column.}

\item{time}{A string giving the column identifying dates.}

\item{data}{A data frame containing the variables used for estimation.}

\item{gvar}{String giving the column that indicates first treatment time for
each state. Use either this option or the combination of \code{treated_states}
and \code{treatment_times}.}

\item{treated_states}{Character values specifying the treated state(s).}

\item{treatment_times}{Specify the treated_states using strings, numbers,
or Dates, corresponding to \code{treated_states}.}

\item{date_format}{Optional string specifying the input date format
when dates are supplied as character strings. Applies to \code{start_date},
\code{end_date}, \code{treatment_times} and the data in the \code{time} column if any
of those are strings.}

\item{covariates}{Optional string or vector of strings specifying
covariates to include.}

\item{ref}{Optional named list indicating the reference category for
categorical covariates.}

\item{ccc}{A string specifying the DID-INT specification.
Any combination of \code{"none"}, \code{"hom"}, \code{"time"}, \code{"state"}, \code{"add"},
and \code{"int"}. Or, alternatively, \code{"all"} (default).}

\item{event}{A logical value used to specify if event study plots should be
made (\code{TRUE}) or if parallel trends plots should be made (\code{FALSE}).}

\item{weights}{A logical value,  if \code{TRUE}, estimates for the event study
plot are computed as weighted averages of state level means for each period
relative to their treatment period; if \code{FALSE}, uses unweighted averages.}

\item{ci}{A number between 0 and 1 used to specify the size of the confidence
bands.}

\item{freq}{Optional string specifying the period length for staggered
adoption. One of \code{"year"}, \code{"month"}, \code{"week"}, \code{"day"}.}

\item{freq_multiplier}{Integer multiplier for \code{freq}. Default is 1.}

\item{start_date}{Optional earliest date to retain in the data.}

\item{end_date}{Optional latest date to retain in the data.}

\item{hc}{Heteroskedasticity-consistent covariance matrix estimator.
One of \code{"hc0"}, \code{"hc1"}, \code{"hc2"}, \code{"hc3"}, \code{"hc4"}.}
}
\value{
A DiDIntPlotObj.
}
\description{
\code{didint_plot()} produces either event study plots or parallel trends plots
depending on what is specified via the \code{event} argument. The parallel trends
plots, as well as the event study plots, are created using the means
residualized by covariates under different model specifications that account
for different violations of the common causal covaraites (CCC) assumptions.
}
\details{
The arguments \code{treated_states} and \code{treatment_times} must be supplied such
that their ordering corresponds with one another. That is, the first
element of \code{treated_states} refers to the state treated at the
date given by the first element of \code{treatment_times}, and so on.

Dates can be entered as strings, numbers, or Date objects.
When character strings are supplied, the input format must be
specified via the \code{date_format} argument (e.g. \code{"yyyy-mm-dd"}).

Period grids are constructed automatically, based on the inputted data
Otherwise, the period grid can be created manually using the arguments
\code{freq}, \code{freq_multiplier}, \code{start_date}, \code{end_date}. More information
on this process can be seen on the didintrjl documentation site:
https://ebjamieson97.github.io/didintrjl/.
}
\examples{
file_path <- system.file("extdata", "merit.csv", package = "didintrjl")
df <- utils::read.csv(file_path)
#didint_plot()

}
\references{
Karim & Webb (2025).
\emph{Good Controls Gone Bad: Difference-in-Differences with Covariates}.
\url{https://arxiv.org/abs/2412.14447}
}
