% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/didint_plot.r
\name{plot.DiDIntPlotObj}
\alias{plot.DiDIntPlotObj}
\title{Plot method for \code{DiDIntPlotObj}}
\usage{
\method{plot}{DiDIntPlotObj}(x, y = NULL, ccc = "all", ...)
}
\arguments{
\item{x}{A \code{DiDIntPlotObj} object.}

\item{y}{\code{NULL} value passed to \code{plot()} method.}

\item{ccc}{Specify which \code{ccc} options you would like plotted from the data.
Any combination of \code{"none"}, \code{"hom"}, \code{"time"}, \code{"state"}, \code{"add"},
and \code{"int"}. Or, alternatively, \code{"all"} (default).}

\item{...}{other arguments}
}
\description{
Plot method for \code{DiDIntPlotObj}
}
