% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/didint.r
\name{print.DiDIntObj}
\alias{print.DiDIntObj}
\title{Print method for \code{DiDIntObj}}
\usage{
\method{print}{DiDIntObj}(x, level = c("agg", "sub"), ...)
}
\arguments{
\item{x}{A \code{DiDIntObj} object.}

\item{level}{Specify either \code{"agg"} or \code{"sub"} to view the aggregate
or sub-aggregate results.}

\item{...}{other arguments}
}
\description{
Print method for \code{DiDIntObj}
}
