% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/didint.r
\name{summary.DiDIntObj}
\alias{summary.DiDIntObj}
\title{Summary method for \code{DiDIntObj}}
\usage{
\method{summary}{DiDIntObj}(object, level = c("agg", "sub"), ...)
}
\arguments{
\item{object}{A \code{DiDIntObj} object}

\item{level}{Specify either \code{"agg"} or \code{"sub"} to view the aggregate
or sub-aggregate results.}

\item{...}{other arguments}
}
\description{
Summary method for \code{DiDIntObj}
}
