% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/didint.r
\name{didint}
\alias{didint}
\title{Estimate the ATT using DID-INT}
\usage{
didint(
  outcome,
  state,
  time,
  data,
  gvar = NULL,
  treated_states = NULL,
  treatment_times = NULL,
  date_format = NULL,
  covariates = NULL,
  ccc = "int",
  agg = "cohort",
  weighting = "both",
  ref = NULL,
  freq = NULL,
  freq_multiplier = 1,
  start_date = NULL,
  end_date = NULL,
  nperm = 999,
  verbose = TRUE,
  seed = sample.int(1e+06, 1),
  notyet = NULL,
  hc = "hc3"
)
}
\arguments{
\item{outcome}{A string giving the column name of the outcome variable.}

\item{state}{A string giving the column identifying states. The state column
should be a character column.}

\item{time}{A string giving the column identifying dates.}

\item{data}{A data frame containing the variables used for estimation.}

\item{gvar}{String giving the column that indicates first treatment time for
each state. Use either this option or the combination of \code{treated_states}
and \code{treatment_times}.}

\item{treated_states}{Character values specifying the treated state(s).}

\item{treatment_times}{Specify the treated_states using strings, numbers,
or Dates, corresponding to \code{treated_states}.}

\item{date_format}{Optional string specifying the input date format
when dates are supplied as character strings. Applies to \code{start_date},
\code{end_date}, \code{treatment_times} and the data in the \code{time} column if any
of those are strings.}

\item{covariates}{Optional string or vector of strings specifying
covariates to include.}

\item{ccc}{A string specifying the DID-INT specification.
One of \code{"hom"}, \code{"time"}, \code{"state"}, \code{"add"}, or \code{"int"} (default \code{"int"}).}

\item{agg}{A string indicating the aggregation method.
One of \code{"cohort"}, \code{"simple"}, \code{"state"}, \code{"sgt"}, \code{"time"} or \code{"none"}.}

\item{weighting}{Weighting scheme to use.
One of \code{"both"}, \code{"att"}, \code{"diff"}, or \code{"none"}.}

\item{ref}{Optional named list indicating the reference category for
categorical covariates.}

\item{freq}{Optional string specifying the period length for staggered
adoption. One of \code{"year"}, \code{"month"}, \code{"week"}, \code{"day"}.}

\item{freq_multiplier}{Integer multiplier for \code{freq}. Default is 1.}

\item{start_date}{Optional earliest date to retain in the data.}

\item{end_date}{Optional latest date to retain in the data.}

\item{nperm}{Number of permutations for randomization inference.
Default is 999.}

\item{verbose}{Logical value, if \code{TRUE}, prints progress during
randomization inference procedure.}

\item{seed}{Integer seed for randomization inference.}

\item{notyet}{Logical value if \code{TRUE}, uses pre-treatment periods from
treated states as controls.}

\item{hc}{Heteroskedasticity-consistent covariance matrix estimator.
One of \code{"hc0"}, \code{"hc1"}, \code{"hc2"}, \code{"hc3"}, \code{"hc4"}.}
}
\value{
Results. ATT, p-values etc.
}
\description{
\code{didint()} estimates the average effect of treatment on the treated (ATT)
using intersection difference-in-differences developped by Karim & Webb
(2025). The method adjusts for covariates that may vary across states,
over time, or jointly by state and time. This function is an R wrapper
around the Julia implementation provided in the \pkg{Didint.jl} package.
For all the details visit the didintrjl documentation site:
https://ebjamieson97.github.io/didintrjl/.
}
\details{
The arguments \code{treated_states} and \code{treatment_times} must be supplied such
that their ordering corresponds with one another. That is, the first
element of \code{treated_states} refers to the state treated at the
date given by the first element of \code{treatment_times}, and so on.

Dates can be entered as strings, numbers, or Date objects.
When character strings are supplied, the input format must be
specified via the \code{date_format} argument (e.g. \code{"yyyy-mm-dd"}).

Period grids for staggered adoption are constructed automatically,
based on the inputted data. Otherwise, the period grid can be created
manually using the arguments \code{freq}, \code{freq_multiplier}, \code{start_date},
and \code{end_date}. More information on this process can be seen on the didintrjl
documentation site: https://ebjamieson97.github.io/didintrjl/.
}
\examples{
file_path <- system.file("extdata", "merit.csv", package = "didintrjl")
df <- utils::read.csv(file_path)
#didint()

}
\references{
Karim & Webb (2025).
\emph{Good Controls Gone Bad: Difference-in-Differences with Covariates}.
\url{https://arxiv.org/abs/2412.14447}

MacKinnon & Webb (2020).
\emph{Randomization inference for difference-in-differences with few treated clusters}.
\doi{10.1016/j.jeconom.2020.04.024}
}
