% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/didint.r
\name{summary.DiDIntObj}
\alias{summary.DiDIntObj}
\title{Summary method for \code{DiDIntObj}}
\usage{
\method{summary}{DiDIntObj}(object, level = c("all", "agg", "sub"), ...)
}
\arguments{
\item{object}{A \code{DiDIntObj} object}

\item{level}{Specify either \code{"agg"}, \code{"sub"}, or \code{"all"}, to view the
results at the aggregate level, the sub-aggregate level, or to view
both simultaneously.}

\item{...}{other arguments}
}
\description{
Summary method for \code{DiDIntObj}
}
